separate product card from its content and embed product editor into it
